csvutil: add NewDecoderFromCSVReader (#18)
* csvutil: add NewDecoderFromCSVReader

The current interface for creating csv decoders (NewDecoder) doesn't
allow you to change the options of the underlying CSV reader. In some
cases, this is necessary (e.g. reading a tab-delimited file). This
change adds a second function NewDecoderFromCSVReader to allow the user
to pass in the CSV reader to use.

* csvutil: add NewEncoderFromCSVWriter